pyproject.toml
[project]
name = "gxp-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dill>=0.4.0",
    "docx2txt>=0.9",
    "faiss-cpu>=1.11.0",
    "flashrank>=0.2.10",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-ollama>=0.3.3",
    "langchain-openai>=0.3.17",
    "langchain-text-splitters>=0.3.8",
    "loguru>=0.7.3",
    "ollama>=0.4.8",
    "omegaconf>=2.3.0",
    "pydantic>=2.11.4",
    "rank-bm25>=0.2.2",
    "ruff>=0.11.10",
    "streamlit>=1.45.1",
    "tqdm>=4.67.1",
    "unstructured>=0.17.2",
]
